authentication_added

+++ src/controller/department.controller.ts
@@ -7,6 +7,10 @@ import CreateDepartmentDto from "../dto/createdepartment.dto";
 import { plainToInstance } from "class-transformer";
 import { validate } from "class-validator";
 import UpdateDepartmentDto from "../dto/updatedepartment.dto";
+import EmployeeService from "../service/employee.service";
+import EmployeeRepository from "../repository/employee.repository";
+import EmployeeController from "./employee.controller";
+import Employee from "../entity/employee.entity";
 
 class DepartmentController {
   public router: express.Router;
@@ -54,36 +58,33 @@ class DepartmentController {
   public createDepartment = async (
     req: express.Request,
     res: express.Response,
-    next:express.NextFunction
+    next: express.NextFunction
   ) => {
-    try{
+    try {
       const departmentdto = plainToInstance(CreateDepartmentDto, req.body);
       const errors = await validate(departmentdto);
       if (errors.length) {
         console.log(JSON.stringify(errors));
         throw new HttpException(400, JSON.stringify(errors));
-      
       }
 
-    const department = await this.departmentService.createDepartment(
-      departmentdto.dept_name,
-      departmentdto.dept_head
-    );
+      const department = await this.departmentService.createDepartment(
+        departmentdto.dept_name,
+        departmentdto.dept_head
+      );
 
-    res.status(201).send(department);
-    }catch(err){
-      next(err)
+      res.status(201).send(department);
+    } catch (err) {
+      next(err);
     }
-    
   };
 
   public updateDepartment = async (
     req: express.Request,
     res: express.Response,
-    next:express.NextFunction
+    next: express.NextFunction
   ) => {
-
-    try{
+    try {
       const updateDto = plainToInstance(UpdateDepartmentDto, req.body);
       const errors = await validate(updateDto);
       if (errors.length) {
@@ -95,23 +96,35 @@ class DepartmentController {
         updateDto.dept_name,
         updateDto.dept_head
       );
-  
-      res.status(201).send(department);
-    }
-    catch(err){
-      next(err)
-    }
 
-    
+      res.status(201).send(department);
+    } catch (err) {
+      next(err);
+    }
   };
 
   public deleteDepartment = async (
     req: express.Request,
-    res: express.Response
+    res: express.Response,
+    next: express.NextFunction
   ) => {
-    const department = await this.departmentService.deleteDepartment(
-      Number(req.params.id)
-    );
+    try {
+      const dept = await this.departmentService.getDepartmentById(
+        Number(req.params.id)
+      );
+      if (!dept) {
+        const error = new HttpException(404, "no department with that id");
+        throw error;
+      }
+      if (dept.employees.length != 0) {
+        throw new HttpException(404, "Dept not empty, cannot delete");
+      }
+      const department = await this.departmentService.deleteDepartment(
+        Number(req.params.id)
+      );
+    } catch (err) {
+      next(err);
+    }
   };
 }
